refactor: isolate verified update responsibilities
